07: Improve API interceptor
